Move database testing code into new testing.rs files
