+++ src/databases/ftl/mod.rs
@@ -8,35 +8,11 @@
 // This file is copyright under the latest version of the EUPL.
 // Please see LICENSE file for your rights under this license.
 
-#[cfg(test)]
-use crate::databases::{
-    common::start_test_transaction, foreign_key_connection::SqliteFKConnectionManager
-};
-#[cfg(test)]
-use diesel::{r2d2::Pool, SqliteConnection};
-
 mod model;
 mod schema;
+#[cfg(test)]
+mod testing;
 
+#[cfg(test)]
+pub use self::testing::*;
 pub use self::{model::*, schema::*};
-
-#[cfg(test)]
-pub const TEST_FTL_DATABASE_PATH: &str = "test/FTL.db";
-
-#[cfg(test)]
-lazy_static! {
-    /// A connection pool for tests which need a database connection
-    static ref CONNECTION_POOL: Pool<SqliteFKConnectionManager> = {
-        let manager = SqliteFKConnectionManager::new(TEST_FTL_DATABASE_PATH);
-        diesel::r2d2::Pool::builder().build(manager).unwrap()
-    };
-}
-
-/// Connect to the testing database
-#[cfg(test)]
-pub fn connect_to_ftl_test_db() -> FtlDatabase {
-    let db = FtlDatabase(CONNECTION_POOL.get().unwrap());
-    start_test_transaction(&db as &SqliteConnection);
-
-    db
-}
